Get requestgroupheader from lib
